// server/src/app/utils/seedSuperAdmin.ts
import { envVars } from "../config/env"
import { IAuthProvider, IUser, Role } from "../modules/user/user.interface"
import { User } from "../modules/user/user.model"
import bcrypt from "bcryptjs";

export const seedSuperAdmin = async () => {
    try {
        const isExistSuperAdmin = await User.findOne({ email: envVars.SUPER_ADMIN_EMAIL })

        if (isExistSuperAdmin) {
            console.log('Super Admin already exist')
            return
        }

        const hashedPassword = await bcrypt.hash(envVars.SUPER_ADMIN_PASSWORD, Number(envVars.BCRYPT_sALT))

        const authProvider: IAuthProvider = { provider: "credentials", providerId: envVars.SUPER_ADMIN_EMAIL }

        const payload: IUser = {
            name: 'Super Admin',
            email: envVars.SUPER_ADMIN_EMAIL,
            password: hashedPassword,
            role: Role.SUPER_ADMIN,
            auths: [authProvider],
            isVerified: true
        }

        const superAdmin = await User.create(payload)
        console.log('Super Admin created successfully')
        console.log(superAdmin)

    } catch (error) {
        console.log(error)
    }
}